+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
-## [Unresolved]
+## [Unreleased]
 ### Added
 - Validators for specified types:
   - Undefined
@@ -21,6 +21,8 @@ for further information about branching and tagging conventions.
   - XDR enum type
   - Creator details
 - Check for spaces in `BaseOperation.isValidString`
+- `DocumentsManager` class to encapsulate creating document's config & 
+  uploading a file to the storage
 
 ### Changed
 - api-caller method `deleteWithSignature` params