+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
-## [Unreleased]
+## [1.7.0-rc.0] - 2019-05-21
 ### Added
 - Validators for specified types:
   - Undefined
@@ -21,7 +21,7 @@ for further information about branching and tagging conventions.
   - XDR enum type
   - Creator details
 - Check for spaces in `BaseOperation.isValidString`
-- `DocumentsManager` class to encapsulate creating document's config & 
+- `DocumentsManager` class to encapsulate creating document's config &
   uploading a file to the storage
 
 ### Changed
@@ -140,7 +140,8 @@ for further information about branching and tagging conventions.
 
 ## [1.3.0] - 2019-03-01
 
-[Unreleased]: https://github.com/tokend/new-js-sdk/compare/1.7.0-x.2...HEAD
+[Unreleased]: https://github.com/tokend/new-js-sdk/compare/1.7.0-rc.0...HEAD
+[1.7.0-rc.0]: https://github.com/tokend/new-js-sdk/compare/1.7.0-x.2...1.7.0-rc.0
 [1.7.0-x.2]: https://github.com/tokend/new-js-sdk/compare/1.7.0-x.1...1.7.0-x.2
 [1.7.0-x.1]: https://github.com/tokend/new-js-sdk/compare/1.7.0-x.0...1.7.0-x.1
 [1.7.0-x.0]: https://github.com/tokend/new-js-sdk/compare/1.6.0...1.7.0-x.0